requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.23.8
cohere==4.37
chromadb==0.4.22
openai==1.12.0  # Upgraded version
httpx==0.25.2   # Added for openai compatibility
langchain==0.0.340
langchain-community==0.0.1
requests==2.31.0
pytesseract==0.3.10
Pillow==10.0.1 